Show only promoted events on frontpage
We have to go with this configuration because of the hackish date filter
added to show events spaning multiple days. It relies on a certain order
of views filters and disabling this, will mess up this order and as a
consequence no events will be shown at all on front.