Revert "PHP 7.2 Compat: Add HTML mail patch"
The patch will not apply because the parameter type is not present in
our lower version of the module, but it is in the patch. We could:

- Update the module to a point where the patch can be applied.
- Make our own patch without the parameter type.

Looking at the discussion in:

https://www.drupal.org/node/3038330

it doesn't seem like the compatability issue will give us any fatal
errors in production and will only have a minor effect in the admin UI.
With this is mind it's decided, that it's not worth the trouble making
this patch apply and we just revert it for now.