Disable and uninstall save_redirect_on_delete module
This module was causing a lot issues when a path was deleted and
was also causing some duplicate entry errors on redirect table in
some situations.

We needed a module that would redirect to a page when a node is
deleted. Hooking into path alias deletion is not the way to go
about this.

We still have the unpublish_node_redirect module, which is working
much better. So we just have to NOT delete content.