Fix missing ting_reference JS on custom cover view
Appearently views doesn't like when we attach JS to the root
element in field_formatter_view(). If we attach it to the $delta
of the added element, it seems work everywhere also when the field
is used in a view.

It doesn't matter that we attach script several times. It will
only get added once.