2173: Add length setting to field formatters
The length settings is now used in minimal view mode, and the
minimal formatters has been removed as it was mostly the same
code.

This also fixes switch statements in
hook_field_formatter_settings_form() and
hook_field_formatter_settings_summary().

Before it was using the field type to compare and not the
formatter type. This could result in formatters getting the
wrong settings form if additional field formatters was added.